+++ src/components/common/Interests/types.d.ts
@@ -1,0 +1,4 @@
+export interface InterestsProps {
+  type: 'cahoot' | 'market';
+  scrollRef: React.RefObject<HTMLDivElement>;
+}
